requirements.txt
# Python dependencies for AI Sample Chopper

# Audio processing
librosa==0.10.1
numpy==1.24.3
soundfile==0.12.1
pydub==0.25.1

# Deep learning
torch==2.0.1
torchaudio==2.0.2
tensorflow==2.12.0

# Web scraping
requests==2.30.0
playwright==1.34.0

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0

# Development tools
pytest==7.3.1
black==23.3.0
